move gates
